backend/requirements.txt
alembic
amqp
annotated-types
anyio
attrs
Authlib
babel
bcrypt
billiard
blinker
boto3
botocore
Brotli
cachelib
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cryptography
cssselect2
Deprecated
distlib
dnspython
email_validator
exceptiongroup
filelock
Flask
Flask-Bcrypt
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-Session
Flask-SQLAlchemy
Flask-WTF
fonttools
google-auth
google-auth-oauthlib
greenlet
h11
html5lib
idna
importlib_metadata
itsdangerous
Jinja2
jmespath
jsonschema
jsonschema-specifications
kombu
limits
Mako
markdown-it-py
MarkupSafe
mdurl
msgspec
mysql-connector-python
numpy
oauthlib
opencv-python
ordered-set
packaging
pandas
pathlib
pillow
platformdirs
prompt_toolkit
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydyf
Pygments
PyJWT
pyphen
python-dateutil
python-dotenv
python-multipart
pytz
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
s3transfer
six
slugify
sniffio
SQLAlchemy
SQLAlchemy-Utils
tinycss2
tinyhtml5
typing_extensions
tzdata
urllib3
uvicorn
vendor
vine
wcwidth
weasyprint
webencodings
Werkzeug
wrapt
WTForms
zipp
zopfli